backend/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
bcrypt==3.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.3
dnspython==2.8.0
easydict==1.13
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.125.0
flatbuffers==25.9.23
fonttools==4.61.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
onnx==1.20.0
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
passlib==1.7.4
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
reportlab==4.4.6
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
simsimd==6.5.8
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
stringzilla==4.5.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.12.12
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
resend==2.4.0
